+++ src/main/java/tr/edu/ogu/ceng/shopingcart/dto/ProductDto.java
@@ -31,11 +31,11 @@ public class ProductDto {
 
     private Category category;
 
-    private Integer createdBy;
+    private Long createdBy;
 
-    private Integer updatedBy;
+    private Long updatedBy;
 
-    private Integer deletedBy;
+    private Long deletedBy;
 
     private LocalDateTime deletedAt;
