+++ src/main/java/tr/edu/ogu/ceng/shopingcart/dto/UserDto.java
@@ -25,11 +25,11 @@ public class UserDto {
 
     private LocalDateTime updatedAt;
 
-    private Integer createdBy;
+    private Long createdBy;
 
-    private Integer updatedBy;
+    private Long updatedBy;
 
-    private Integer deletedBy;
+    private Long deletedBy;
 
     private LocalDateTime deletedAt;
